backend/requirements.txt
# HealthIQ Backend Dependencies
django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
pymongo>=4.6
python-dotenv>=1.0
